pull

+++ app/Http/Controllers/SupplierController.php
@@ -7,34 +7,56 @@ use App\Http\Requests\UpdateSupplierRequest;
 use App\Models\Brand;
 use App\Models\Category;
 use App\Models\Supplier;
+use CloudinaryLabs\CloudinaryLaravel\Facades\Cloudinary;
 use Illuminate\Http\Request;
 use Faker\Provider\Image;
 
 class SupplierController extends Controller
 {
-    protected function uploadSupplierImage($request)
-    {
-        $supplierImage = $request->file('photo');
-        $imageType = $supplierImage->getClientOriginalExtension();
-        $imageName = rand(100, 100000) . $request->name . '.' . $imageType;
-        $directory = 'inventory/supplier-images/';
-        $imageUrl = $directory . $imageName;
-        Image::make($supplierImage)->save($imageUrl);
-        return $imageUrl;
-
-    }
+    //    protected function uploadSupplierImage($request)
+//    {
+//        $supplierImage = $request->file('photo');
+//        $imageType = $supplierImage->getClientOriginalExtension();
+//        $imageName = rand(100, 100000) . $request->name . '.' . $imageType;
+//        $directory = 'inventory/supplier-images/';
+//        $imageUrl = $directory . $imageName;
+//        Image::make($supplierImage)->save($imageUrl);
+//        return $imageUrl;
+//
+//    }
 
     public function saveSupplier(Request $request)
     {
         $supplier = new Supplier();
 
-        // upload image to storage
+        // upload image to storage for local
+//        if ($request->hasFile('photo')) {
+//            $image = $request->file('photo');
+//            $image_name = time() . '.' . $image->getClientOriginalExtension();
+//            $image->move(public_path('images/suppliers/'), $image_name);
+//            $supplier->photo = $image_name;
+//        } else if ($request->photo) {
+//            $base64_string = $request->photo;
+//            $image = base64_decode($base64_string);
+//
+//            $file_name = time() . '.' . 'png';
+//            $file_path = public_path('images/suppliers/' . $file_name);
+//            file_put_contents($file_path, $image);
+//            $supplier->photo = $file_name;
+//        }
+
+        // upload image to storage for Cloudinary
         if ($request->hasFile('photo')) {
-            $image = $request->file('photo');
-            $image_name = time() . '.' . $image->getClientOriginalExtension();
-            $image->move(public_path('images'), $image_name);
-            $supplier->photo = $image_name;
+            $imagePath = $request->file('photo')->getRealPath();
+            $uploadedImage = Cloudinary::upload($imagePath)->getSecurePath();
+            $supplier->photo = $uploadedImage;
+        } else if ($request->photo) {
+            $imageBase64 = $request->photo;
+            $uploadedImage = Cloudinary::upload("data:image/png;base64," . $imageBase64)->getSecurePath();
+            $supplier->photo = $uploadedImage;
         }
+
+
         $supplier->name = $request->name;
         $supplier->email = $request->email;
         $supplier->phone = $request->phone;
@@ -84,17 +106,57 @@ class SupplierController extends Controller
         $data['bank_number'] = $request->bank_number;
 
 
+        // if ($request->hasFile('photo')) {
+        //     $image = $request->file('photo');
+        //     $image_name = time() . '.' . $image->getClientOriginalExtension();
+        //     $image->move(public_path('images/suppliers/'), $image_name);
+        //     $data['photo'] = $image_name; // remove old image
+        //     $old_image = public_path('images/suppliers/' . $supplier->photo);
+        //     if (file_exists($old_image)) {
+        //         @unlink($old_image);
+        //     }
+        // } else if ($request->photo) {
+        //     $base64_string = $request->photo;
+        //     $image = base64_decode($base64_string);
+
+        //     $file_name = time() . '.' . 'png';
+        //     $file_path = public_path('images/suppliers/' . $file_name);
+        //     file_put_contents($file_path, $image);
+        //     $data['photo'] = $file_name; // remove old image
+        //     $old_image = public_path('images/suppliers/' . $supplier->photo);
+        //     if (file_exists($old_image)) {
+        //         @unlink($old_image);
+        //     }
+        // }
+
+
+        // upload image to storage for Cloudinary
         if ($request->hasFile('photo')) {
-            $image = $request->file('photo');
-            $image_name = time() . '.' . $image->getClientOriginalExtension();
-            $image->move(public_path('images'), $image_name);
-            $data['photo'] = $image_name;
-            // remove old image
-            $old_image = public_path('images/' . $supplier->photo);
-            if (file_exists($old_image)) {
-                @unlink($old_image);
+
+            //check if old image exist
+            $old_image = Cloudinary::getImage($supplier->photo);
+            if ($old_image->getPublicId() != null) {
+                Cloudinary::destroy($old_image->getPublicId());
             }
+
+            //set image to response json
+            $imagePath = $request->file('photo')->getRealPath();
+            $uploadedImage = Cloudinary::upload($imagePath)->getSecurePath();
+            $data['photo'] = $uploadedImage;
+
+
+        } else if ($request->photo) {
+            //check if old image exist
+            $old_image = Cloudinary::getImage($supplier->photo);
+            if ($old_image->getPublicId() != null) {
+                Cloudinary::destroy($old_image->getPublicId());
+            }
+            //set image to response json
+            $imageBase64 = $request->photo;
+            $uploadedImage = Cloudinary::upload("data:image/png;base64," . $imageBase64)->getSecurePath();
+            $data['photo'] = $uploadedImage;
         }
+
         $supplier->update($data);
         return response()->json([
             "message" => "Supplier data updated successfully!",
@@ -107,8 +169,20 @@ class SupplierController extends Controller
     {
         $supplier = Supplier::findOrFail($id);
 
-        //if($supplier->photo) unlink($supplier->photo);
-        $supplier->delete();
+        if ($supplier->products->isEmpty()) {
+            // do something if there are no products
+            $supplier->delete();
+            //check if old image exist
+            $old_image = Cloudinary::getImage($supplier->photo);
+            if ($old_image->getPublicId() != null) {
+                Cloudinary::destroy($old_image->getPublicId());
+            }
+        } else {
+            // do something if there are products
+            return response()->json([
+                "message" => "Products exist!",
+            ], 404);
+        }
         return response()->json([
             "message" => "Supplier removed successfully!",
             "status" => 200,
